+++ app/mailers/user_mailer.rb
@@ -11,6 +11,14 @@ class UserMailer < ActionMailer::Base
          reply_to: user.email)
   end
 
+  def dua_submitted(system_admin, agreement)
+    setup_email
+    @system_admin = system_admin
+    @agreement = agreement
+    mail(to: system_admin.email,
+         subject: "#{agreement.user.name} Submitted a Data Use Agreement")
+  end
+
   protected
 
   def setup_email