+++ test/controllers/user_mailer_test.rb
@@ -16,4 +16,18 @@ class UserMailerTest < ActionMailer::TestCase
     assert_match(/#{valid.name} \[#{valid.email}\] has signed up for an account\./, email.encoded)
   end
 
+  test "dua submitted email" do
+    agreement = agreements(:one)
+    admin = users(:admin)
+
+    # Send the email, then test that it got queued
+    email = UserMailer.dua_submitted(admin, agreement).deliver
+    assert !ActionMailer::Base.deliveries.empty?
+
+    # Test the body of the sent email contains what we expect it to
+    assert_equal [admin.email], email.to
+    assert_equal "#{agreement.user.name} Submitted a Data Use Agreement", email.subject
+    assert_match(/#{agreement.user.name} \[#{agreement.user.email}\] has submitted a Data Use Agreement\./, email.encoded)
+  end
+
 end